requirements.txt
# Honeypot Security System - Dependencies
# Python 3.8+ required

# Web Framework
flask>=2.0.0
werkzeug>=2.0.0

# SSH Honeypot
paramiko>=3.0.0
cryptography>=41.0.0

# Configuration
pyyaml>=6.0

# HTTP Requests (Discord webhook)
requests>=2.28.0

# GeoIP Lookup
geoip2>=4.0.0
maxminddb>=2.0.0

# Database
# SQLite is built-in with Python

# Process Management
python-daemon>=2.3.0; sys_platform != 'win32'

# Utilities
python-dateutil>=2.8.0
